Refactor to split up the components in MilestonesBoard
